Create README.md
Getting Started with React1Copyright 2023 Code with Moshcodewithmosh.comGetting StartedComponentsJavaScript FrameworkJavaScript LibraryJSXDOMVirtual DOM TermsSummary•React is a JavaScript library for building dynamic and interactive user interfaces.•In React applications, we don’t query and update the DOM. Instead, we describe our application using small, reusable components. React will take care of efficiently creating and updating DOM elements.•React components can be created using a function or a class. Function-based components are the preferred approach as they’re more concise and easier to work with.•JSX stands for JavaScript XML. It is a syntax that allows us to write components that combine HTML and JavaScript in a readable and expressive way, making it easier to create complex user interfaces.•When our application starts, React takes a tree of components and builds a JavaScript data structure called the virtual DOM. This virtual DOM is different from the actual DOM in the browser. It’s a lightweight, in-memory representation of our component tree.